requirements.txt
openai
PyMuPDF
faiss-cpu
numpy
nltk